requirements.txt
fastapi
uvicorn
python-multipart
PyMuPDF
pdfplumber
sentence-transformers
faiss-cpu
python-jose